% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ettbc-package.R
\docType{package}
\name{ettbc-package}
\alias{ettbc}
\alias{ettbc-package}
\title{ettbc: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ucdavis/ettbc}
  \item \url{https://ucdavis.github.io/ettbc/}
  \item Report bugs at \url{https://github.com/ucdavis/ettbc/issues}
}

}
\author{
\strong{Maintainer}: Douglas Ezra Morrison \email{demorrison@ucdavis.edu} (\href{https://orcid.org/0000-0002-7195-830X}{ORCID})

}
