% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_function.R
\name{example_function}
\alias{example_function}
\title{Example Function}
\usage{
example_function(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The median of the input vector
}
\description{
This is an example function that demonstrates basic functionality.
}
\examples{
example_function(c(1, 2, 3, 4, 5))
}
