% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpt-package.R
\docType{package}
\name{rpt-package}
\alias{rpt}
\alias{rpt-package}
\title{rpt: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UCD-SERG/rpt}
  \item \url{https://ucd-serg.github.io/rpt/}
  \item Report bugs at \url{https://github.com/UCD-SERG/rpt/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
