% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ettbc-package.R
\docType{package}
\name{ettbc-package}
\alias{ettbc}
\alias{ettbc-package}
\title{ettbc: Emulating a Target Trial for Breast Cancer Screening}
\description{
Implements the clone-censor step of the target trial emulation methodology for studying the effect of sustained screening mammography on breast cancer mortality, following the approach of García-Albéniz et al. (2020) \doi{10.7326/M18-1199}. Provides functions to clone participants into two trial arms (STOPBASE and CONTINUE), apply per-arm censoring rules, and expand the resulting dataset to a long format suitable for inverse-probability-weighted survival analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ucdavis/ettbc}
  \item \url{https://ucdavis.github.io/ettbc/}
  \item Report bugs at \url{https://github.com/ucdavis/ettbc/issues}
}

}
\author{
\strong{Maintainer}: Douglas Ezra Morrison \email{demorrison@ucdavis.edu} (\href{https://orcid.org/0000-0002-7195-830X}{ORCID})

Authors:
\itemize{
  \item Douglas Ezra Morrison \email{demorrison@ucdavis.edu} (\href{https://orcid.org/0000-0002-7195-830X}{ORCID})
}

}
